--- a/app/src/main/java/bg/zahov/app/MainNavGraph.kt
+++ b/app/src/main/java/bg/zahov/app/MainNavGraph.kt
@@ -4,7 +4,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Modifier
 import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavGraph.Companion.findStartDestination
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
@@ -25,8 +24,8 @@ import bg.zahov.app.ui.settings.profile.EditProfileScreen
 import bg.zahov.app.ui.welcome.WelcomeScreen
 import bg.zahov.app.ui.workout.AddTemplateWorkoutScreen
 import bg.zahov.app.ui.workout.WorkoutScreen
-import bg.zahov.app.ui.workout.rest.RestScreen
 import bg.zahov.app.ui.workout.finish.WorkoutFinishScreen
+import bg.zahov.app.ui.workout.rest.RestScreen
 import bg.zahov.app.ui.workout.start.StartWorkoutScreen
 
 @Composable
@@ -127,7 +126,8 @@ fun MainNavGraph(
         composable<Exercises> {
             ExercisesScreen(
                 navigateInfo = { navController.navigate(ExerciseInfo) },
-                navigateBack = { navController.navigateUp() })
+                navigateBack = {
+                    navController.navigateUp() })
         }
         composable<ExerciseAdd> {
             AddExerciseScreen(