--- a/app/src/main/java/bg/zahov/app/ui/BottomBar.kt
+++ b/app/src/main/java/bg/zahov/app/ui/BottomBar.kt
@@ -9,9 +9,14 @@ import androidx.compose.material3.NavigationBarItemDefaults
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
+import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.semantics.semantics
+import androidx.compose.ui.semantics.testTag
+import androidx.compose.ui.semantics.testTagsAsResourceId
 import androidx.navigation.NavDestination.Companion.hasRoute
 import androidx.navigation.NavDestination.Companion.hierarchy
 import androidx.navigation.NavGraph.Companion.findStartDestination
@@ -24,6 +29,7 @@ import bg.zahov.app.Measure
 import bg.zahov.app.StartWorkout
 import bg.zahov.fitness.app.R
 
+@OptIn(ExperimentalComposeUiApi::class)
 @SuppressLint("RestrictedApi")
 @Composable
 fun BottomBar(
@@ -70,6 +76,11 @@ fun BottomBar(
         ) {
             screens.forEach { screen ->
                 NavigationBarItem(
+                    modifier = Modifier
+                        .semantics {
+                            testTagsAsResourceId = true
+                        }
+                        .testTag(stringResource(screen.titleId)),
                     label = {
                         Text(text = stringResource(screen.titleId))
                     },