--- a/app/src/main/java/bg/zahov/app/ui/custom/CommonPasswordField.kt
+++ b/app/src/main/java/bg/zahov/app/ui/custom/CommonPasswordField.kt
@@ -13,11 +13,15 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
+import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.Shape
+import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.semantics.semantics
+import androidx.compose.ui.semantics.testTagsAsResourceId
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.unit.dp
@@ -25,6 +29,7 @@ import androidx.compose.ui.unit.sp
 import bg.zahov.fitness.app.R
 import kotlinx.coroutines.delay
 
+@OptIn(ExperimentalComposeUiApi::class)
 @Composable
 fun CommonPasswordField(
     modifier: Modifier = Modifier,
@@ -46,13 +51,14 @@ fun CommonPasswordField(
     ),
     onPasswordChange: (String) -> Unit,
     onPasswordVisibilityChange: (Boolean) -> Unit,
-    enabled: Boolean = true
+    enabled: Boolean = true,
+    testTag: String = ""
 ) {
     var isError by remember { mutableStateOf(false) }
     var isActive by remember { mutableStateOf(false) }
 
     TextField(
-        modifier = modifier,
+        modifier = modifier.semantics { testTagsAsResourceId = true }.testTag(testTag),
         value = password,
         onValueChange = {
             onPasswordChange(it)