… app/baselineprofile/src/main/java/com/example/baselineprofile/FinishWorkoutJourney.kt
… app/src/main/java/bg/zahov/app/ui/exercise/ExerciseViewModel.kt
… app/src/main/java/bg/zahov/app/ui/exercise/ExercisesScreen.kt
… app/src/main/java/bg/zahov/app/ui/history/HistoryScreen.kt
… app/src/main/java/bg/zahov/app/ui/history/calendar/CalendarScreen.kt
… app/src/main/java/bg/zahov/app/ui/history/info/topbar/HistoryInfoTopBar.kt
… app/src/main/java/bg/zahov/app/ui/settings/SettingsScreen.kt
… app/src/main/java/bg/zahov/app/ui/topbar/TopBarAction.kt
… app/src/main/java/bg/zahov/app/ui/topbar/TopBarBack.kt
… app/src/main/java/bg/zahov/app/ui/topbar/workout/TopBarWorkout.kt
… app/src/main/java/bg/zahov/app/ui/workout/WorkoutScreen.kt
… app/src/main/java/bg/zahov/app/ui/workout/start/StartWorkoutScreen.kt
… app/src/release/generated/baselineProfiles/baseline-prof.txt
… app/src/release/generated/baselineProfiles/startup-prof.txt