Trello -> [https://trello.com/c/omswoMCO/103-finish-workout-journey](url)


**Startup Time** Improvement:
**Partial With Baseline Profiles:** 5.85% faster than Partial Compilation (503.9 ms to 474.4 ms).
**No Compilation:** 1.89% faster than Partial Compilation (503.9 ms to 494.4 ms).

**Results**
![image](https://github.com/user-attachments/assets/53b22c94-2872-4c35-84f8-e04291a23368)